requirements.txt
python-docx
openpyxl
Pillow
Babel